Fix toggle all in filtered views
