+++ src/TodoApp.js
@@ -1,5 +1,4 @@
 import React, { useState, useEffect, useMemo, useContext } from 'react';
-import { useLocation } from 'react-router-dom';
 
 import { DispatchContext, TodosContext } from './context/TodosContext';
 import { actions } from './context/reducer';
@@ -7,38 +6,13 @@ import { NewTodo } from './components/NewTodo';
 import { TodoList } from './components/TodoList';
 import { TodoFooter } from './components/TodoFooter';
 import { UserInfo } from './components/UserInfo';
-import { FILTERS, USER_ID } from './constants';
+import { USER_ID } from './constants';
 import { getUser, getUserTodos, deleteTodo } from './api';
 
 function TodoApp() {
   const todos = useContext(TodosContext);
   const dispatch = useContext(DispatchContext);
   const [user, setUser] = useState({});
-
-  const { pathname } = useLocation();
-
-  const filteredTodos = useMemo(
-    () => (
-      todos.filter(({ completed }) => {
-        switch (pathname) {
-          case FILTERS.active:
-            return !completed;
-
-          case FILTERS.completed:
-            return completed;
-
-          case FILTERS.all:
-          default:
-            return true;
-        }
-      })
-    )
-    , [todos, pathname],
-  );
-
-  const toggleAllChecked = useMemo(() => (
-    todos.every(({ completed }) => completed)
-  ), [todos]);
 
   useEffect(() => {
     getUser(USER_ID)
@@ -77,10 +51,7 @@ function TodoApp() {
           <NewTodo />
         </header>
 
-        <TodoList
-          todos={filteredTodos}
-          toggleAllChecked={toggleAllChecked}
-        />
+        <TodoList />
 
         {todos.length > 0 && (
           <TodoFooter
